The Shootake subsystem and commands
This is a really base version of our shootake subsystem and commands where I have created the subsystem, the shoot command, the intake command, and the default command. I have also added different constants and added the subsystem to robot container. The actual motor speeds and can ids will not be correct however and have numbers as placeholders for later. Furthermore, the commands will need to be updated for sensor use.