More Intake and shoot commands
Created different commands for shooting in the amp and speaker. I also modified the intake command for when the shintake goes under the arm. I also modified the default command to use a similar intake system without letting it shoot out. Plus the variable intake now uses a similar system.